Extract Search command into separate assembly with module support
Move UnitySearchHandler to its own asmdef (UniCli.Server.Editor.Search)
following the same plugin pattern as NuGet and BuildMagic. Tag with
[Module("Search")] so it can be disabled via the module system.
