+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Search/UniCli.Server.Editor.Search.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "UniCli.Server.Editor.Search",
+    "rootNamespace": "UniCli.Server.Editor.Handlers.Search",
+    "references": [
+        "UniCli.Server.Editor"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
