+++ src/UniCli.Unity/Packages/com.yucchiy.unicli-server/Editor/Handlers/Search/UniCli.Server.Editor.Search.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bb8cb4bc48f7e4a289bd61bd1d4c8499
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
